+++ java/src/test/java/io/cucumber/java/defaultstransformer/DataTableSteps.java
@@ -1,6 +1,6 @@
 package io.cucumber.java.defaultstransformer;
 
-import io.cucumber.datatable.dependency.com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.cucumber.java.DefaultDataTableCellTransformer;
 import io.cucumber.java.DefaultDataTableEntryTransformer;
 import io.cucumber.java.DefaultParameterTransformer;